requirements.txt
Scrapy~=2.11.2

requests~=2.32.3
mysql-connector-python~=9.1.0
redis~=5.0.4
twisted~=24.7.0